started refactoring the app, and got tests working
