+++ app/__init__.py
@@ -1,0 +1,24 @@
+from flask import Flask, render_template
+from flask_bootstrap import Bootstrap
+from flask_sqlalchemy import SQLAlchemy
+from config import config
+
+bootstrap = Bootstrap()
+db = SQLAlchemy()
+
+def create_app(config_name):
+    app = Flask(__name__, template_folder='templates')
+    print("config_name:%s" % (config_name,))
+    config_class = config[config_name]
+    app.config.from_object(config_class)
+    config_class.init_app(app)
+
+    bootstrap.init_app(app)
+    db.init_app(app)
+
+    # TODO: attach routes and error pages here
+
+    from .main import main as main_blueprint
+    app.register_blueprint(main_blueprint)
+
+    return app
