+++ app/templates/customers.html
