+++ info/NOTES.txt
@@ -62,7 +62,7 @@ $ docker push morspin/socksess:v0.3
 Ensure app.yaml image refers to docker.io/morspin/socksess:v0.4.x
 2. docker build -t morspin/socksess .
 3. docker tag morspin/socksess:latest morspin/socksess:v0.4.x
-4. docker push morspin/socksess:v0.4.x
+4. docker push morspin/socksess: v0.4.x
 5. k apply -f k8s/app.yaml -n socksess
 6. watch k get pods,services -n socksess
 7. copy the socksess pod name
@@ -140,9 +140,11 @@ $ add the command to run the migration-upgrade to `scripts/run.sh`
 0.5.1 - first try with migrations, sqlite
 0.5.2 - first try with migrations, postgresql
 0.5.3 - ^, specify production mode to get postgres, not sqlite
-0.5.4 - type "datetime" does not exist - try sd.TIMESTAMP
+0.5.4 - type "datetime" does not exist - try sd.TIMESTAMP -- works
 
 
 2. Add blueprints (chapter 7) 
+
+0.5.5 - first try with blueprints
 
 3. Add API endpoints (chapter 14)
