+++ k8s/app.yaml
@@ -16,7 +16,7 @@ spec:
     spec:
       containers:
       - name: socksessm
-        image: docker.io/morspin/socksess:v0.5.4
+        image: docker.io/morspin/socksess:v0.5.5
         imagePullPolicy: Always
         ports:
         - containerPort: 5000
